requirements.txt
Flask==2.3.3
pyrebase4==4.7.1
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.0.0
python-dateutil==2.8.2
Werkzeug==2.3.7
requests
